requirements.txt
snowflake-connector-python
grpcio==1.67.1
python-dotenv
sqlalchemy
snowflake-sqlalchemy
pandas
streamlit
groq